-Implemented Maths for cost and volume of pipe -Implemented valid check method within pipe -Created pipe type 2 -Created Console Pipe creation -Edited UI to include table --include valid method within each update method?

+++ PipesRus/src/PipeII.java
@@ -5,7 +5,7 @@ public class PipeII extends Pipe
 
   public PipeII(int plastic,ArrayList<String> colours, Boolean insulated,
                     Boolean reinforced, Boolean chemicalResist,
-                    float length,float outerDiameter)
+                    double length,double outerDiameter)
   {
     super(plastic, colours, insulated,reinforced, chemicalResist, length, outerDiameter);
     valid();
